+++ backend/NumberInformation.API/NumberInformation.Tests/PrimeServiceTest.cs
@@ -18,16 +18,16 @@ namespace NumberInformation.Tests
             _primeService = new ();
         }
 
-        [TestCase(new int[] { 1, 2, 4, 5, 10, 20, 25, 50, 100 }, new int[] { 1, 2, 5})]
-        [TestCase(new int[] { 1, 13 }, new int[] { 1, 13 })]
-        [TestCase(new int[] { 1, 2, 3, 4, 6, 8, 12, 24 }, new int[] { 1, 2, 3 })]
-        [TestCase(new int[] { 1, 999983 }, new int[] { 1, 999983 })]
+        [TestCase(new long[] { 1, 2, 4, 5, 10, 20, 25, 50, 100 }, new long[] { 1, 2, 5})]
+        [TestCase(new long[] { 1, 13 }, new long[] { 1, 13 })]
+        [TestCase(new long[] { 1, 2, 3, 4, 6, 8, 12, 24 }, new long[] { 1, 2, 3 })]
+        [TestCase(new long[] { 1, 999983 }, new long[] { 1, 999983 })]
         
-        public void PrimeNumbers_TestCase(int[] dividingNumbers, int[] primeNumbers)
+        public void PrimeNumbers_TestCase(long[] dividingNumbers, long[] primeNumbers)
         {
             var result = dividingNumbers.Where(_primeService.IsPrime).ToList();
 
-            result.ShouldBe(new List<int>(primeNumbers));
+            result.ShouldBe(new List<long>(primeNumbers));
         }
 
     }
